Updated contract.
Mint one of one reserved function now counts towards max supply, so that
was changed to 5000.

setTokenURI calls weren't adding the tokenID to the uri, so they were
all the same. Now fixed.

Removed a space from the voucher hash function, maybe that was breaking
it? anyways the vouchers work now as expected which is nice.